release cache RAM as soon as it is no longer needed
* clean up middle layer before output layer, not the other way round
* free worker thread clones right afer pending relations have been processed

So far the cache was only freed at the very and, and so the RAM
allocated by it was not available to PostgreSQL as maintenance_work_mem
during the index building stage unless the now unused cache got
swapped out.

By freeing the cache early and returning the RAM allocated to it to the
operating system it is possible to set both the osm2pgsql cache size
and maintenance_work_mem to the same size (close to the free RAM
size) without swapping as index building only happens after the
cache has been freed, while previously both parameters had to
be tuned carefully to not use more than the free RAM size in total
to avoid swapping.